backend/requirements.txt
fastapi
uvicorn
requests
transformers
numpy
scikit-learn
google-generativeai
torch
pydantic
sentencepiece